+++ CHANGELOG.md
@@ -10,7 +10,8 @@
 - Dropped `isNode` function in favor of `instanceof Node` notation.
 - `cached` decorator now supports properties as well as methods.
 - Dropped exposed `Node` methods that asumed a sibling order.
-- Added methods in `Literal` node to check value type. 
+- Added methods in `Literal` node to check value type.
+- Dropped `defaultFieldValues` in favor of the cleaner `defaultValueFor` method.
 
 ## v3.1.1
 - Fixed bug where closures where unable to retain self reference when nested inside other closures.